+++ changelog.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
-## [unreleased]
+## [0.3.0] - 2022-04-16
 
+- Fix issue with last experience rendered as oneliner
 - Replace options -t, --template with flags -e, --experience
 - Empty toyear defaults to current year